requirements.txt
requests
python-dotenv
PyYAML
mkdocs
mkdocs-material
